backport patch to fix SIGABRT when closing
already merged upstream

upstream made the behavior opt-in for backward compatibility reasons, but it's always the right thing to do.

This applies the option unconditionally for a smaller patch.

Consequences of prior behavior: chance of SIGABRT if any context is closed or key issued while CURVE is in use.
Consequences of new behavior: up to one open FD on /dev/urandom, user may need/want to to close